Logic to handle node values charts
