+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-06-12 01:04:35 +0100 using RuboCop version 0.67.2.
+# on 2019-07-09 16:54:18 +0100 using RuboCop version 0.67.2.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -22,7 +22,7 @@ Layout/ClosingParenthesisIndentation:
   Exclude:
     - 'app/services/api/v3/import/stats.rb'
 
-# Offense count: 37
+# Offense count: 39
 # Cop supports --auto-correct.
 Layout/EmptyLineAfterGuardClause:
   Enabled: false
@@ -48,6 +48,24 @@ Layout/EndAlignment:
 Layout/IndentArray:
   EnforcedStyle: consistent
 
+# Offense count: 1
+# Cop supports --auto-correct.
+# Configuration parameters: EnforcedStyle, EnforcedStyleForEmptyBraces.
+# SupportedStyles: space, no_space
+# SupportedStylesForEmptyBraces: space, no_space
+Layout/SpaceBeforeBlockBraces:
+  Exclude:
+    - 'spec/services/api/v3/manage_download_attributes_spec.rb'
+
+# Offense count: 4
+# Cop supports --auto-correct.
+# Configuration parameters: EnforcedStyle, EnforcedStyleForEmptyBraces.
+# SupportedStyles: space, no_space, compact
+# SupportedStylesForEmptyBraces: space, no_space
+Layout/SpaceInsideHashLiteralBraces:
+  Exclude:
+    - 'spec/controllers/admin/top_profiles_controller_spec.rb'
+
 # Offense count: 2
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle.
@@ -56,6 +74,11 @@ Layout/TrailingBlankLines:
   Exclude:
     - 'Rakefile'
     - 'config/deploy/staging.rb'
+
+# Offense count: 2
+Lint/AmbiguousBlockAssociation:
+  Exclude:
+    - 'spec/services/api/v3/manage_download_attributes_spec.rb'
 
 # Offense count: 1
 # Configuration parameters: AllowSafeAssignment.
@@ -68,32 +91,32 @@ Lint/ShadowingOuterLocalVariable:
   Exclude:
     - 'lib/tasks/charts.rake'
 
-# Offense count: 2
+# Offense count: 1
 # Cop supports --auto-correct.
 # Configuration parameters: IgnoreEmptyBlocks, AllowUnusedKeywordArguments.
 Lint/UnusedBlockArgument:
   Exclude:
-    - 'app/services/concerns/api/v3/dashboards/charts/helpers.rb'
+    - 'cms_ideas/batch_save_form.rb'
 
 # Offense count: 1
 Lint/UselessAssignment:
   Exclude:
     - 'lib/tasks/db_remote_import.rake'
 
-# Offense count: 64
+# Offense count: 65
 Metrics/AbcSize:
-  Max: 37
+  Max: 31
 
-# Offense count: 13
+# Offense count: 14
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
   Max: 288
 
-# Offense count: 5
+# Offense count: 6
 Metrics/CyclomaticComplexity:
   Max: 9
 
-# Offense count: 160
+# Offense count: 169
 # Configuration parameters: CountComments, ExcludedMethods.
 Metrics/MethodLength:
   Max: 62
@@ -101,9 +124,9 @@ Metrics/MethodLength:
 # Offense count: 2
 # Configuration parameters: CountComments.
 Metrics/ModuleLength:
-  Max: 208
+  Max: 159
 
-# Offense count: 8
+# Offense count: 9
 Metrics/PerceivedComplexity:
   Max: 10
 
@@ -111,6 +134,17 @@ Metrics/PerceivedComplexity:
 Naming/AccessorMethodName:
   Exclude:
     - 'lib/db_helpers/search_path_helpers.rb'
+
+# Offense count: 1
+# Configuration parameters: NamePrefix, NamePrefixBlacklist, NameWhitelist, MethodDefinitionMacros.
+# NamePrefix: is_, has_, have_
+# NamePrefixBlacklist: is_, has_, have_
+# NameWhitelist: is_a?
+# MethodDefinitionMacros: define_method, define_singleton_method
+Naming/PredicateName:
+  Exclude:
+    - 'spec/**/*'
+    - 'app/models/concerns/api/v3/is_downloadable.rb'
 
 # Offense count: 13
 # Configuration parameters: MinNameLength, AllowNamesEndingInNumbers, AllowedNames, ForbiddenNames.
@@ -158,23 +192,12 @@ Style/IfUnlessModifier:
     - 'app/services/api/v3/flows/filter.rb'
 
 # Offense count: 1
-Style/MultilineBlockChain:
-  Exclude:
-    - 'app/services/api/v3/dashboards/parametrised_charts.rb'
-
-# Offense count: 1
 # Cop supports --auto-correct.
 # Configuration parameters: EnforcedStyle, MinBodyLength.
 # SupportedStyles: skip_modifier_ifs, always
 Style/Next:
   Exclude:
     - 'lib/tasks/charts.rake'
-
-# Offense count: 2
-# Cop supports --auto-correct.
-Style/PerlBackrefs:
-  Exclude:
-    - 'app/services/concerns/api/v3/dashboards/charts/helpers.rb'
 
 # Offense count: 1
 # Cop supports --auto-correct.
