+++ app/controllers/api/v3/dashboards/charts/multi_year_ncont_overview_controller.rb
@@ -3,7 +3,7 @@ module Api
     module Dashboards
       module Charts
         class MultiYearNcontOverviewController < ApiController
-          include FilterParams
+          include ChartParams
           skip_before_action :load_context
 
           def index
@@ -15,7 +15,7 @@ module Api
             ensure_required_param_present(:ncont_attribute_id)
 
             render json: MultiYearNcontOverview.new(
-              ChartParameters.new(chart_params)
+              ChartParameters::FlowValues.new(flow_values_chart_params)
             ).call
           end
         end
