+++ app/controllers/api/v3/dashboards/charts/multi_year_node_values_overview_controller.rb
@@ -2,11 +2,9 @@ module Api
   module V3
     module Dashboards
       module Charts
-        class MultiYearNoNcontNodeTypeViewController < ApiController
+        class MultiYearNodeValuesOverviewController < ApiController
           include ChartParams
           skip_before_action :load_context
-
-          DEFAULT_TOP_N = 20
 
           def index
             ensure_required_param_present(:country_id)
@@ -14,11 +12,10 @@ module Api
             ensure_required_param_present(:start_year)
             ensure_required_param_present(:end_year)
             ensure_required_param_present(:cont_attribute_id)
-            ensure_required_param_present(:node_type_id)
-            ensure_required_param_set(:top_n, DEFAULT_TOP_N)
+            ensure_required_param_present(:node_id)
 
-            render json: MultiYearNoNcontNodeTypeView.new(
-              ChartParameters::FlowValues.new(flow_values_chart_params)
+            render json: MultiYearNodeValuesOverview.new(
+              ChartParameters::NodeValues.new(node_values_chart_params)
             ).call
           end
         end
