+++ app/controllers/api/v3/dashboards/parametrised_charts_controller.rb
@@ -2,7 +2,7 @@ module Api
   module V3
     module Dashboards
       class ParametrisedChartsController < ApiController
-        include FilterParams
+        include ChartParams
         skip_before_action :load_context
 
         def index
@@ -10,11 +10,18 @@ module Api
           ensure_required_param_present(:commodity_id)
           ensure_required_param_present(:cont_attribute_id)
 
-          render json: ParametrisedCharts.new(ChartParameters.new(chart_params)).call,
+          parametrised_charts = ParametrisedCharts::List.new(
+            ChartParameters::FlowValues.new(flow_values_chart_params)
+          ).call
+
+          render json: parametrised_charts,
                  root: 'data',
                  each_serializer: Api::V3::Dashboards::ParametrisedChartSerializer,
                  url: proc { |options|
-                   send(:"api_v3_dashboards_charts_#{options.delete(:source)}_index_url", options)
+                   send(
+                     :"api_v3_dashboards_charts_#{options.delete(:source)}_index_url",
+                     options
+                   )
                  }
         end
       end
