+++ app/models/api/v3/context.rb
@@ -30,6 +30,7 @@ module Api
       has_many :contextual_layers
       has_many :context_node_types
       has_many :profiles, through: :context_node_types
+      has_many :dashboards_attributes
       has_many :download_attributes
       has_many :map_attribute_groups
       has_many :map_attributes, through: :map_attribute_groups
