+++ app/services/api/v3/dashboards/charts/multi_year_no_ncont_node_type_view.rb
@@ -5,10 +5,11 @@ module Api
       module Charts
         class MultiYearNoNcontNodeTypeView
           include Api::V3::Dashboards::Charts::Helpers
+          include Api::V3::Dashboards::Charts::FlowValuesHelpers
 
           OTHER = 'OTHER'.freeze
 
-          # @param chart_parameters [Api::V3::Dashboards::ChartParameters]
+          # @param chart_parameters [Api::V3::Dashboards::ChartParameters::FlowValues]
           def initialize(chart_parameters)
             @chart_parameters = chart_parameters
             @cont_attribute = chart_parameters.cont_attribute
