+++ app/services/api/v3/dashboards/charts/multi_year_node_values_overview.rb
@@ -1,17 +1,18 @@
-# Data for bar chart widget
+# Data for dynamic sentence widget
 module Api
   module V3
     module Dashboards
       module Charts
-        class MultiYearNoNcontOverview
+        class MultiYearNodeValuesOverview
           include Api::V3::Dashboards::Charts::Helpers
-          include Api::V3::Dashboards::Charts::FlowValuesHelpers
+          include Api::V3::Dashboards::Charts::NodeValuesHelpers
 
-          # @param chart_parameters [Api::V3::Dashboards::ChartParameters::FlowValues]
+          # @param chart_parameters [Api::V3::Dashboards::ChartParameters::NodeValues]
           def initialize(chart_parameters)
             @chart_parameters = chart_parameters
             @cont_attribute = chart_parameters.cont_attribute
             @context = chart_parameters.context
+            @node = chart_parameters.node
             @start_year = chart_parameters.start_year
             @end_year = chart_parameters.end_year
             initialize_query
@@ -34,11 +35,13 @@ module Api
           private
 
           def initialize_query
-            @query = flow_query
+            @query = @cont_attribute.original_attribute.
+              node_values.
+              where(node_id: @node.id).
+              select('SUM(value) AS y0').
+              group(:node_id)
             apply_year_x
-            apply_cont_attribute_y
             apply_multi_year_filter
-            apply_flow_path_filters
           end
         end
       end
