+++ app/services/api/v3/dashboards/charts/single_year_ncont_overview.rb
@@ -5,8 +5,9 @@ module Api
       module Charts
         class SingleYearNcontOverview
           include Api::V3::Dashboards::Charts::Helpers
+          include Api::V3::Dashboards::Charts::FlowValuesHelpers
 
-          # @param chart_parameters [Api::V3::Dashboards::ChartParameters]
+          # @param chart_parameters [Api::V3::Dashboards::ChartParameters::FlowValues]
           def initialize(chart_parameters)
             @chart_parameters = chart_parameters
             @cont_attribute = chart_parameters.cont_attribute
