+++ app/services/api/v3/dashboards/charts/single_year_node_values_overview.rb
@@ -1,30 +1,30 @@
-# Data for bar chart widget
+# Data for dynamic sentence widget
 module Api
   module V3
     module Dashboards
       module Charts
-        class MultiYearNoNcontOverview
+        class SingleYearNodeValuesOverview
           include Api::V3::Dashboards::Charts::Helpers
-          include Api::V3::Dashboards::Charts::FlowValuesHelpers
+          include Api::V3::Dashboards::Charts::NodeValuesHelpers
 
-          # @param chart_parameters [Api::V3::Dashboards::ChartParameters::FlowValues]
+          # @param chart_parameters [Api::V3::Dashboards::ChartParameters::NodeValues]
           def initialize(chart_parameters)
             @chart_parameters = chart_parameters
             @cont_attribute = chart_parameters.cont_attribute
             @context = chart_parameters.context
-            @start_year = chart_parameters.start_year
-            @end_year = chart_parameters.end_year
+            @node = chart_parameters.node
+            @year = chart_parameters.start_year
             initialize_query
           end
 
           def call
             @data = @query.map do |r|
-              r.attributes.slice('x', 'y0').symbolize_keys
+              r.attributes.slice('y0').symbolize_keys
             end
             @meta = {
-              xAxis: year_axis_meta,
+              xAxis: {},
               yAxis: axis_meta(@cont_attribute, 'number'),
-              x: year_legend_meta,
+              x: {},
               y0: legend_meta(@cont_attribute),
               info: info
             }
@@ -34,11 +34,12 @@ module Api
           private
 
           def initialize_query
-            @query = flow_query
-            apply_year_x
-            apply_cont_attribute_y
-            apply_multi_year_filter
-            apply_flow_path_filters
+            @query = @cont_attribute.original_attribute.
+              node_values.
+              where(node_id: @node.id).
+              select('SUM(value) AS y0').
+              group(:node_id, :year)
+            apply_single_year_filter if temporal?
           end
         end
       end
