+++ spec/services/api/v3/dashboards/charts/multi_year_ncont_overview_spec.rb
@@ -27,7 +27,7 @@ RSpec.describe Api::V3::Dashboards::Charts::MultiYearNcontOverview do
   }
 
   let(:chart_parameters) {
-    Api::V3::Dashboards::ChartParameters.new(parameters_hash)
+    Api::V3::Dashboards::ChartParameters::FlowValues.new(parameters_hash)
   }
 
   let(:result) {
