+++ spec/services/api/v3/dashboards/charts/multi_year_no_ncont_node_type_view_spec.rb
@@ -71,7 +71,7 @@ RSpec.describe Api::V3::Dashboards::Charts::MultiYearNoNcontNodeTypeView do
   }
 
   let(:chart_parameters) {
-    Api::V3::Dashboards::ChartParameters.new(parameters_hash)
+    Api::V3::Dashboards::ChartParameters::FlowValues.new(parameters_hash)
   }
 
   let(:result) {
