+++ spec/services/api/v3/dashboards/charts/single_year_ncont_node_type_view_spec.rb
@@ -29,7 +29,7 @@ RSpec.describe Api::V3::Dashboards::Charts::SingleYearNcontNodeTypeView do
   }
 
   let(:chart_parameters) {
-    Api::V3::Dashboards::ChartParameters.new(parameters_hash)
+    Api::V3::Dashboards::ChartParameters::FlowValues.new(parameters_hash)
   }
 
   let(:result) {
