+++ spec/services/api/v3/dashboards/charts/single_year_no_ncont_overview_spec.rb
@@ -21,7 +21,7 @@ RSpec.describe Api::V3::Dashboards::Charts::SingleYearNoNcontOverview do
   }
 
   let(:chart_parameters) {
-    Api::V3::Dashboards::ChartParameters.new(parameters_hash)
+    Api::V3::Dashboards::ChartParameters::FlowValues.new(parameters_hash)
   }
 
   let(:result) {
