+++ spec/services/api/v3/dashboards/parametrised_charts/flow_values_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
+RSpec.describe Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts do
   include_context 'api v3 brazil context node types'
   include_context 'api v3 brazil recolor by attributes'
   include_context 'api v3 brazil resize by attributes'
@@ -32,8 +32,8 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
   let(:multi_year) { {start_year: 2016, end_year: 2017} }
 
   let(:chart_types) {
-    Api::V3::Dashboards::ParametrisedCharts.new(
-      Api::V3::Dashboards::ChartParameters.new(parameters)
+    Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts.new(
+      Api::V3::Dashboards::ChartParameters::FlowValues.new(parameters)
     ).call
   }
 
@@ -66,7 +66,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       [
         {
           source: :single_year_no_ncont_overview,
-          type: Api::V3::Dashboards::ParametrisedCharts::DYNAMIC_SENTENCE,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::DYNAMIC_SENTENCE,
           x: nil
         }
       ] + [
@@ -79,7 +79,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       ].map do |node_type|
         {
           source: :single_year_no_ncont_node_type_view,
-          type: Api::V3::Dashboards::ParametrisedCharts::HORIZONTAL_BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::HORIZONTAL_BAR_CHART,
           x: :node_type,
           node_type_id: node_type.id
         }
@@ -100,7 +100,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       [
         {
           source: :multi_year_no_ncont_overview,
-          type: Api::V3::Dashboards::ParametrisedCharts::BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::BAR_CHART,
           x: :year
         }
       ] + [
@@ -113,7 +113,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       ].map do |node_type|
         {
           source: :multi_year_no_ncont_node_type_view,
-          type: Api::V3::Dashboards::ParametrisedCharts::STACKED_BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::STACKED_BAR_CHART,
           x: :year,
           break_by: :node_type,
           node_type_id: node_type.id
@@ -135,7 +135,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       [
         {
           source: :single_year_ncont_overview,
-          type: Api::V3::Dashboards::ParametrisedCharts::DONUT_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::DONUT_CHART,
           x: :ncont_attribute
         }
       ] + [
@@ -148,7 +148,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       ].map do |node_type|
         {
           source: :single_year_ncont_node_type_view,
-          type: Api::V3::Dashboards::ParametrisedCharts::HORIZONTAL_STACKED_BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::HORIZONTAL_STACKED_BAR_CHART,
           x: :node_type,
           break_by: :ncont_attribute,
           node_type_id: node_type.id
@@ -170,7 +170,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       [
         {
           source: :multi_year_ncont_overview,
-          type: Api::V3::Dashboards::ParametrisedCharts::STACKED_BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::STACKED_BAR_CHART,
           x: :year,
           break_by: :ncont_attribute
         }
@@ -184,7 +184,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       ].map do |node_type|
         {
           source: :multi_year_no_ncont_node_type_view,
-          type: Api::V3::Dashboards::ParametrisedCharts::STACKED_BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::STACKED_BAR_CHART,
           x: :year,
           break_by: :node_type,
           node_type_id: node_type.id
@@ -216,13 +216,13 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       [
         {
           source: :multi_year_ncont_overview,
-          type: Api::V3::Dashboards::ParametrisedCharts::STACKED_BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::STACKED_BAR_CHART,
           x: :year,
           break_by: :ncont_attribute
         },
         {
           source: :multi_year_ncont_overview,
-          type: Api::V3::Dashboards::ParametrisedCharts::STACKED_BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::STACKED_BAR_CHART,
           x: :year,
           break_by: :ncont_attribute,
           node_type_id: api_v3_exporter_node_type.id,
@@ -231,7 +231,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
         },
         {
           source: :multi_year_ncont_overview,
-          type: Api::V3::Dashboards::ParametrisedCharts::STACKED_BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::STACKED_BAR_CHART,
           x: :year,
           break_by: :ncont_attribute,
           node_type_id: api_v3_exporter_node_type.id,
@@ -247,7 +247,7 @@ RSpec.describe Api::V3::Dashboards::ParametrisedCharts do
       ].map do |node_type|
         {
           source: :multi_year_no_ncont_node_type_view,
-          type: Api::V3::Dashboards::ParametrisedCharts::STACKED_BAR_CHART,
+          type: Api::V3::Dashboards::ParametrisedCharts::FlowValuesCharts::STACKED_BAR_CHART,
           x: :year,
           break_by: :node_type,
           node_type_id: node_type.id
