+++ spec/support/schemas/dashboards_charts_single_year_node_values_overview.json
@@ -1,0 +1,213 @@
+{
+  "definitions": {},
+  "$schema": "http://json-schema.org/draft-06/schema#",
+  "$id": "http://example.com/root.json",
+  "type": "object",
+  "required": [
+    "data",
+    "meta"
+  ],
+  "properties": {
+    "data": {
+      "$id": "#/properties/data",
+      "type": "array",
+      "items": {
+        "$id": "#/properties/data/items",
+        "type": "object",
+        "required": [
+          "y0"
+        ],
+        "properties": {
+          "y0": {
+            "$id": "#/properties/data/items/properties/y0",
+            "type": "number"
+          }
+        }
+      }
+    },
+    "meta": {
+      "$id": "#/properties/meta",
+      "type": "object",
+      "required": [
+        "xAxis",
+        "yAxis",
+        "x",
+        "y0",
+        "info"
+      ],
+      "properties": {
+        "xAxis": {
+          "$id": "#/properties/meta/properties/xAxis",
+          "type": "object"
+        },
+        "yAxis": {
+          "$id": "#/properties/meta/properties/yAxis",
+          "type": "object",
+          "required": [
+            "type",
+            "label",
+            "prefix",
+            "format",
+            "suffix"
+          ],
+          "properties": {
+            "type": {
+              "$id": "#/properties/meta/properties/yAxis/properties/type",
+              "type": "string",
+              "pattern": "^(.*)$"
+            },
+            "label": {
+              "$id": "#/properties/meta/properties/yAxis/properties/label",
+              "type": "string",
+              "pattern": "^(.*)$"
+            },
+            "prefix": {
+              "$id": "#/properties/meta/properties/yAxis/properties/prefix",
+              "type": "string",
+              "pattern": "^(.*)$"
+            },
+            "format": {
+              "$id": "#/properties/meta/properties/yAxis/properties/format",
+              "type": "string",
+              "pattern": "^(.*)$"
+            },
+            "suffix": {
+              "$id": "#/properties/meta/properties/yAxis/properties/suffix",
+              "type": "string",
+              "pattern": "^(.*)$"
+            }
+          }
+        },
+        "x": {
+          "$id": "#/properties/meta/properties/x",
+          "type": "object"
+        },
+        "y0": {
+          "$id": "#/properties/meta/properties/y0",
+          "type": "object",
+          "required": [
+            "label",
+            "tooltip"
+          ],
+          "properties": {
+            "label": {
+              "$id": "#/properties/meta/properties/y0/properties/label",
+              "type": "string",
+              "pattern": "^(.*)$"
+            },
+            "tooltip": {
+              "$id": "#/properties/meta/properties/y0/properties/tooltip",
+              "type": "object",
+              "required": [
+                "prefix",
+                "format",
+                "suffix"
+              ],
+              "properties": {
+                "prefix": {
+                  "$id": "#/properties/meta/properties/y0/properties/tooltip/properties/prefix",
+                  "type": "string",
+                  "pattern": "^(.*)$"
+                },
+                "format": {
+                  "$id": "#/properties/meta/properties/y0/properties/tooltip/properties/format",
+                  "type": "string",
+                  "pattern": "^(.*)$"
+                },
+                "suffix": {
+                  "$id": "#/properties/meta/properties/y0/properties/tooltip/properties/suffix",
+                  "type": "string",
+                  "pattern": "^(.*)$"
+                }
+              }
+            }
+          }
+        },
+        "info": {
+          "$id": "#/properties/meta/properties/info",
+          "type": "object",
+          "required": [
+            "temporal",
+            "years",
+            "filter"
+          ],
+          "properties": {
+            "temporal": {
+              "$id": "#/properties/meta/properties/info/properties/temporal",
+              "type": "boolean"
+            },
+            "years": {
+              "$id": "#/properties/meta/properties/info/properties/years",
+              "type": "object",
+              "required": [
+                "start_year",
+                "end_year"
+              ],
+              "properties": {
+                "start_year": {
+                  "$id": "#/properties/meta/properties/info/properties/years/properties/start_year",
+                  "type": "integer"
+                },
+                "end_year": {
+                  "$id": "#/properties/meta/properties/info/properties/years/properties/end_year",
+                  "type": "null"
+                }
+              }
+            },
+            "filter": {
+              "$id": "#/properties/meta/properties/info/properties/filter",
+              "type": "object",
+              "required": [
+                "cont_attribute",
+                "node"
+              ],
+              "properties": {
+                "cont_attribute": {
+                  "$id": "#/properties/meta/properties/info/properties/filter/properties/cont_attribute",
+                  "type": "string",
+                  "pattern": "^(.*)$"
+                },
+                "node": {
+                  "$id": "#/properties/meta/properties/info/properties/filter/properties/node",
+                  "type": "object",
+                  "required": [
+                    "id",
+                    "name",
+                    "node_type_id",
+                    "node_type",
+                    "profile"
+                  ],
+                  "properties": {
+                    "id": {
+                      "$id": "#/properties/meta/properties/info/properties/filter/properties/node/properties/id",
+                      "type": "integer"
+                    },
+                    "name": {
+                      "$id": "#/properties/meta/properties/info/properties/filter/properties/node/properties/name",
+                      "type": "string",
+                      "pattern": "^(.*)$"
+                    },
+                    "node_type_id": {
+                      "$id": "#/properties/meta/properties/info/properties/filter/properties/node/properties/node_type_id",
+                      "type": "integer"
+                    },
+                    "node_type": {
+                      "$id": "#/properties/meta/properties/info/properties/filter/properties/node/properties/node_type",
+                      "type": "string",
+                      "pattern": "^(.*)$"
+                    },
+                    "profile": {
+                      "$id": "#/properties/meta/properties/info/properties/filter/properties/node/properties/profile",
+                      "type": "string",
+                      "pattern": "^(.*)$"
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
